+++ src/mimiry/__init__.py
@@ -10,14 +10,17 @@ from mimiry.exceptions import (
     SessionError,
     SessionFailed,
     SessionTimeout,
+    MapError,
     ResultParseError,
     ResultIntegrityError,
 )
+from mimiry._serialization import RemoteFunctionError
+from mimiry._session import RunInfo
 from mimiry.function import function, Function
 from mimiry.image import Image
 from mimiry.run import run
 
-__version__ = "0.3.3"
+__version__ = "0.4.0"
 
 __all__ = [
     "__version__",
@@ -32,6 +35,9 @@ __all__ = [
     "SessionError",
     "SessionFailed",
     "SessionTimeout",
+    "MapError",
     "ResultParseError",
     "ResultIntegrityError",
+    "RemoteFunctionError",
+    "RunInfo",
 ]